[Bugfix] double `next()` in checkAuth middleware
Notes:
The server kept crashing when POST requesting `/login` twice in a row
with valid user credentials both times. The error it threw was
'ERR_HTTP_HEADERS_SENT', which is usually thrown if you try to send
multiple responses from the same endpoint route (i.e. attempting to
write headers after already sending a response).

My guess is that the most common scenario for this bug is when
`res.send/json()` is called once inside a conditional and then
again outisde/below the conditional, yet still in the same endpoint
route. This is fixed relatively easily; simply pre-pend `return` to
the first `res.send/json()` to terminate the route function.

But in my case, the problem was NOT that I ended the req/res cycle
twice inside the `/login` route. The problem was that I called
`next()` twice in my `checkAuth` custom middleware.

My original implementation of `checkAuth` in Reddit.js worked just
fine, but that version was not linted. I took the same code and tried
using it here, except with a linter this time. ESLint was giving me a
'consistent-return' rule error. To satisfy the linter, I added some
return statements. And that's how I unknowingly wrote the bug in the
first place.

Finding this bug's source was incredibly challenging. The stack trace
couldn't detect my double `next()` call. Instead, it pointed me to an
ENTIRELY different location altogether. If I inserted a `res.json()`
right before `await User.findOne(...)` (src/controllers/auth, line 20),
then the server did NOT crash. But if I moved the `res.json()` just ONE
line below that query, it caused the server to start crashing again.

I was under the impression that the Mongoose query, for some
inconceivable reason, was causing the problem. Which made NO sense to
me. I tried refactoring the query countless times, but nothing seemed
to work. I was SO CONFUSED (and sad) all day long. I kept questiong,
"how can it even be possible for a Mongoose query to cause an HTTP
header error?!" In the end, the query had nothing to do with it,
despite its execution being DIRECTLY related to reproducing the issue.

Debugging this took me hours upon hours upon hours. At first, I thought
the presence of the nToken cookie was the origin of the problem. I was
not able to find ANY solution online that helped me fix the code. Most
of what I found regarding the 'ERR_HTTP_HEADERS_SENT' error simply
talked about the double `res.send/json()` issue mentioned above, but no
mention of double `next()` calls in custom middleware.

BY FAR the most challenging debugging session I've experienced so far.
I'm so grateful I made it out! Alhamdulillah.

This is also the longest commit message I've ever written. It looks
more like a blog post than a commit message 😆.